rev: change some parameters
Parameter shift_one controls mainly the load threshold for turning cpu1 online.
When 2 or more cores are online, the load threshold will dynamically change
and add 30(shift_one + 30). Make shift_one now be the control for the load
threshold when more than 2 cores are online and make 20 be the default and non-
configurable threshold when only 1 core is online. Same applies for down_shift.

Signed-off-by: revjunkie <rajibrahim@rocketmail.com>